+++ fabric-content-registries-v0/build.gradle
@@ -1,11 +1,7 @@
 archivesBaseName = "fabric-content-registries-v0"
-version = getSubprojectVersion(project, "0.2.0")
+version = getSubprojectVersion(project, "0.1.9")
 
 dependencies {
 	compile project(path: ':fabric-api-base', configuration: 'dev')
 	compile project(path: ':fabric-resource-loader-v0', configuration: 'dev')
-
-	testmodCompile project(path: ':fabric-object-builder-api-v1', configuration: 'dev')
-	testmodCompile project(path: ':fabric-tool-attribute-api-v1', configuration: 'dev')
-	testmodCompile project(path: ':fabric-tag-extensions-v0', configuration: 'dev')
 }
