+++ fabric-content-registries-v0/src/main/resources/fabric-content-registries-v0.mixins.json
@@ -4,10 +4,7 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "MixinAbstractFurnaceBlockEntity",
-    "MixinFireBlock",
-    "MixinMobEntity",
-    "MixinPlayerEntity",
-    "MixinRemoveOffHandItemTask"
+    "MixinFireBlock"
   ],
   "injectors": {
     "defaultRequire": 1
