--- fabric-content-registries-v0/src/testmod/resources/fabric.mod.json
@@ -1,19 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-content-registries-v0-testmod",
-  "name": "Fabric Content Registries API (v0) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-content-registries-v0": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.shield.ShieldTest"
-    ],
-    "client": [
-      "net.fabricmc.fabric.test.shield.ShieldTestClient"
-    ]
-  }
-}
