+++ fabric-item-api-v1/build.gradle
@@ -1,6 +1,11 @@
 archivesBaseName = "fabric-item-api-v1"
-version = getSubprojectVersion(project, "1.0.0")
+version = getSubprojectVersion(project, "1.1.0")
 
 dependencies {
 	compile project(path: ':fabric-api-base', configuration: 'dev')
+
+	testmodCompile project(path: ':fabric-object-builder-api-v1', configuration: 'dev')
+	testmodCompile project(path: ':fabric-tool-attribute-api-v1', configuration: 'dev')
+	testmodCompile project(path: ':fabric-tag-extensions-v0', configuration: 'dev')
+	testmodCompile project(path: ':fabric-resource-loader-v0', configuration: 'dev')
 }
