+++ fabric-item-api-v1/src/testmod/resources/assets/fabric/lang/en_us.json
