+++ fabric-item-api-v1/src/testmod/resources/assets/fabric/models/item/shield.json
