+++ fabric-item-api-v1/src/testmod/resources/fabric.mod.json
@@ -2,15 +2,19 @@
   "schemaVersion": 1,
   "id": "fabric-item-api-v1-testmod",
   "name": "Fabric Item API (v1) Test Mod",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "environment": "*",
   "license": "Apache-2.0",
   "depends": {
     "fabric-item-api-v1": "*"
   },
   "entrypoints": {
+    "main": [
+      "net.fabricmc.fabric.test.item.ShieldTest"
+    ],
     "client": [
-      "net.fabricmc.fabric.test.item.client.TooltipTests"
+      "net.fabricmc.fabric.test.item.client.TooltipTests",
+      "net.fabricmc.fabric.test.item.client.ShieldTestClient"
     ]
   }
 }
